Refactor project imports
